Break up types.rs into different files
