+++ src/types/config.rs
@@ -1,0 +1,7 @@
+use super::ModuleConfig;
+
+/// Combined RX and TX config
+pub struct Config {
+    pub tx: ModuleConfig,
+    pub rx: ModuleConfig,
+}
